refactor(current-account): split saga_handlers.go into domain-specific files
Extract position-keeping and financial-accounting handler implementations
into saga_handler_position_keeping.go and saga_handler_financial_accounting.go.
Retain registration, shared types, and utilities in saga_handlers.go (388 lines,
down from 1013). Remove from knownOversizedFiles.
