+++ tests/architecture/size_test.go
@@ -27,7 +27,6 @@ const (
 var knownOversizedFiles = map[string]bool{
 	"services/control-plane/internal/applier/grpc_handler.go":               true,
 	"services/control-plane/internal/generator/validate_fix.go":             true,
-	"services/current-account/service/saga_handlers.go":                     true,
 	"services/identity/service/server.go":                                   true,
 	"services/internal-account/service/server.go":                           true,
 	"services/market-information/service/observation_service.go":            true,
